Conflicts:
	main/features/org.talend.tos.libraries.feature/feature.xml

+++ main/plugins/org.talend.repository/src/main/java/org/talend/repository/ui/actions/EditPropertiesAction.java
@@ -14,7 +14,9 @@ package org.talend.repository.ui.actions;
 
 import java.io.File;
 import java.lang.reflect.InvocationTargetException;
+import java.util.List;
 
+import org.apache.commons.lang.StringUtils;
 import org.eclipse.core.resources.IFile;
 import org.eclipse.core.resources.IFolder;
 import org.eclipse.core.resources.IResource;
@@ -61,17 +63,25 @@ import org.talend.core.model.properties.ByteArray;
 import org.talend.core.model.properties.ConnectionItem;
 import org.talend.core.model.properties.Item;
 import org.talend.core.model.properties.JobScriptItem;
+import org.talend.core.model.properties.Property;
 import org.talend.core.model.properties.RoutineItem;
 import org.talend.core.model.properties.SQLPatternItem;
+import org.talend.core.model.relationship.Relation;
+import org.talend.core.model.relationship.RelationshipItemBuilder;
 import org.talend.core.model.repository.ERepositoryObjectType;
 import org.talend.core.model.repository.IRepositoryViewObject;
 import org.talend.core.model.repository.RepositoryManager;
+import org.talend.core.model.routines.CodesJarInfo;
+import org.talend.core.model.routines.RoutinesUtil;
 import org.talend.core.model.update.RepositoryUpdateManager;
 import org.talend.core.repository.ui.editor.RepositoryEditorInput;
 import org.talend.core.runtime.process.ITalendProcessJavaProject;
 import org.talend.core.services.IUIRefresher;
+import org.talend.core.utils.CodesJarResourceCache;
 import org.talend.designer.core.ICamelDesignerCoreService;
 import org.talend.designer.core.IDesignerCoreService;
+import org.talend.designer.maven.tools.CodesJarM2CacheManager;
+import org.talend.designer.maven.utils.CodesJarMavenUtil;
 import org.talend.designer.runprocess.IRunProcessService;
 import org.talend.metadata.managment.ui.wizard.PropertiesWizard;
 import org.talend.metadata.managment.ui.wizard.process.EditProcessPropertiesWizard;
@@ -170,6 +180,27 @@ public class EditPropertiesAction extends AContextualAction {
                 }
             }
 
+            // warn re-generate all pom after codejar rename
+            ERepositoryObjectType objectType = node.getObjectType();
+            if (!originalName.equals(object.getProperty().getLabel())
+                    && ERepositoryObjectType.getAllTypesOfCodesJar().contains(objectType)) {
+                String relationType = null;
+                if (ERepositoryObjectType.ROUTINESJAR != null && ERepositoryObjectType.ROUTINESJAR.equals(objectType)) {
+                    relationType = RelationshipItemBuilder.ROUTINES_JAR_RELATION;
+                } else if (ERepositoryObjectType.BEANSJAR != null && ERepositoryObjectType.BEANSJAR.equals(objectType)) {
+                    relationType = RelationshipItemBuilder.BEANS_JAR_RELATION;
+                }
+                if (StringUtils.isNotBlank(relationType)) {
+                    List<Relation> itemsRelatedTo = RelationshipItemBuilder.getInstance()
+                            .getAllVersionItemsRelatedTo(object.getProperty().getId(), relationType, true);
+                    if (!itemsRelatedTo.isEmpty()) {
+                        MessageDialog.openWarning(Display.getCurrent().getActiveShell(),
+                                Messages.getString("EditPropertiesAction.warning"), //$NON-NLS-1$
+                                Messages.getString("EditPropertiesAction.warnToReGenerateAllPom")); //$NON-NLS-1$
+                    }
+                }
+            }
+
         }
     }
 
@@ -218,7 +249,16 @@ public class EditPropertiesAction extends AContextualAction {
     protected void processRename(IRepositoryNode node, String originalName) {
         try {
             IRunProcessService runProcessService = CorePlugin.getDefault().getRunProcessService();
-            ITalendProcessJavaProject talendProcessJavaProject = runProcessService.getTalendCodeJavaProject(node.getObjectType());
+            ITalendProcessJavaProject talendProcessJavaProject = null;
+            Property property = node.getObject().getProperty();
+            boolean isInnerCode = RoutinesUtil.isInnerCodes(property);
+            CodesJarInfo codeJarinfo = null;
+            if (isInnerCode) {
+                codeJarinfo = CodesJarResourceCache.getCodesJarByInnerCode((RoutineItem) property.getItem());
+                talendProcessJavaProject = runProcessService.getTalendCodesJarJavaProject(codeJarinfo);
+            } else {
+                talendProcessJavaProject = runProcessService.getTalendCodeJavaProject(node.getObjectType());
+            }
             if (talendProcessJavaProject == null) {
                 return;
             }
@@ -226,13 +266,15 @@ public class EditPropertiesAction extends AContextualAction {
             IFolder srcFolder = talendProcessJavaProject.getSrcFolder();
             IPackageFragmentRoot root = talendProcessJavaProject.getJavaProject().getPackageFragmentRoot(srcFolder);
 
-            // add for bug TDI-24379 on August 23, 2013.
-            IFolder srcInterFolder = srcFolder.getFolder(JavaUtils.JAVA_INTERNAL_DIRECTORY);
-            if (srcInterFolder.exists()) {
-                File file = new File(srcInterFolder.getLocationURI());
-                for (File f : file.listFiles()) {
-                    if (f.isFile()) {
-                        f.delete();
+            if (!isInnerCode) {
+                // add for bug TDI-24379 on August 23, 2013.
+                IFolder srcInterFolder = srcFolder.getFolder(JavaUtils.JAVA_INTERNAL_DIRECTORY);
+                if (srcInterFolder.exists()) {
+                    File file = new File(srcInterFolder.getLocationURI());
+                    for (File f : file.listFiles()) {
+                        if (f.isFile()) {
+                            f.delete();
+                        }
                     }
                 }
             }
@@ -299,6 +341,9 @@ public class EditPropertiesAction extends AContextualAction {
             }
             RoutineItem item = (RoutineItem) node.getObject().getProperty().getItem();
             IFile javaFile = (IFile) newUnit.getAdapter(IResource.class);
+            if (javaFile == null || !javaFile.exists()) {
+                return;
+            }
             try {
                 ByteArray byteArray = item.getContent();
                 byteArray.setInnerContentFromFile(javaFile);
@@ -310,6 +355,10 @@ public class EditPropertiesAction extends AContextualAction {
                 ExceptionHandler.process(e);
             }
 
+            if (isInnerCode && codeJarinfo != null) {
+                CodesJarM2CacheManager.updateCodesJarProject(codeJarinfo.getProperty());
+            }
+
         } catch (Exception e) {
             // e.printStackTrace();
             ExceptionHandler.process(e);
@@ -317,8 +366,14 @@ public class EditPropertiesAction extends AContextualAction {
     }
 
     protected IPackageFragment getPackageFragment(IPackageFragmentRoot root, IRepositoryNode node) {
-        String folder = node.getContentType().getFolder();
-        String packageName = Path.fromOSString(folder).lastSegment();
+        String packageName = null;
+        Property property = node.getObject().getProperty();
+        if (RoutinesUtil.isInnerCodes(property)) {
+            packageName = CodesJarMavenUtil.getCodesJarPackageByInnerCode((RoutineItem) property.getItem());
+        } else {
+            String folder = node.getContentType().getFolder();
+            packageName = Path.fromOSString(folder).lastSegment();
+        }
         return root.getPackageFragment(packageName);
     }
 
@@ -362,7 +417,9 @@ public class EditPropertiesAction extends AContextualAction {
                 switch (node.getType()) {
                 case REPOSITORY_ELEMENT:
                     if (node.getObjectType() == ERepositoryObjectType.BUSINESS_PROCESS
-                            || node.getObjectType() == ERepositoryObjectType.PROCESS) {
+                            || node.getObjectType() == ERepositoryObjectType.PROCESS
+                            || node.getObjectType() == ERepositoryObjectType.ROUTINESJAR
+                            || node.getObjectType() == ERepositoryObjectType.BEANSJAR) {
                         canWork = true;
                     } else if (node.getObjectType() == ERepositoryObjectType.ROUTINES) {
                         Item item = node.getObject().getProperty().getItem();
